chore: create a single CLI download image (#227)
* chore: create a single CLI download image

The Konflux team has made changes to the build-source-image task, which
previously failed when adding more than just cosign, gitsign and
rekor-cli to the image. When we added ec to it, the task began to fail
due to an overly large results string, which Konflux could not handle.
That limitation in Konflux has been eliminated.

The change in this commit is to create a single CLI download image,
which can be used by the operator, instead of the multiple currently in
place. Once this change is committed on the main branch, a new component
can be created in Konflux to test the build. If it is successful, we can
remove the following components from the `cli` Application in Konflux,
and subsequently remove those Dockerfiles from this repo.

The following components can be removed after this build is successful.

* cli/client-server-cg
* cli/client-server-ec
* cli/client-server-f

Signed-off-by: Lance Ball <lball@redhat.com>

* fixup: use httpd image and serve the binaries

Instead of having the operator provide the httpd server, include it
here to simplify cluster deployment.

Signed-off-by: Lance Ball <lball@redhat.com>

* fixup: put binaries in /var/www/html/clients

Signed-off-by: Lance Ball <lball@redhat.com>

---------

Signed-off-by: Lance Ball <lball@redhat.com>